Fix CopySubresourceRegion uint args and nullable routing log warning

+++ src/VirtualDofMatrix.App/Rendering/GpuInstancedMatrixRenderer.cs
@@ -975,7 +975,7 @@ public sealed class GpuInstancedMatrixRenderer : IMatrixRenderer
         var srcX = (_surfaceWidth - copyWidth) / 2;
         var srcY = (_surfaceHeight - copyHeight) / 2;
         var srcBox = new Box(srcX, srcY, 0, srcX + copyWidth, srcY + copyHeight, 1);
-        _context.CopySubresourceRegion(_directPresentBackBuffer, 0, dstX, dstY, 0, _gpuCompositeTexture, 0, srcBox);
+        _context.CopySubresourceRegion(_directPresentBackBuffer, 0, (uint)dstX, (uint)dstY, 0, _gpuCompositeTexture, 0, srcBox);
     }
 
     private bool ShouldUseGpuDotPass(DotStyleConfig style)
